+++ steelc/src/building/linking/link_error.h
@@ -1,0 +1,8 @@
+#pragma once
+
+enum link_error {
+	LINK_OK,
+	LINK_FAIL_LOAD_IR,
+	LINK_FAIL_NO_IR,
+	LINK_FAIL_LINK_MODULE,
+};