+++ steelc/src/codegen/codegen.cpp
@@ -3,19 +3,25 @@
 #include <iostream>
 #include <string>
 #include <vector>
+#include <memory>
+
+#include <llvm/IR/LLVMContext.h>
+#include <llvm/IR/Module.h>
 
 #include <ast/compilation_unit.h>
 #include <codegen/codegen_visitor.h>
 #include <codegen/error/codegen_exception.h>
-#include <codegen/ir/ir_holder.h>
+#include <codegen/codegen_result.h>
 #include <output/output.h>
 
-std::vector<ir_holder> codegen::generate_ir() {
-	std::vector<ir_holder> ir_modules;
+std::unique_ptr<codegen_result> codegen::generate_modules() {
+	auto result = std::make_unique<codegen_result>(std::make_unique<llvm::LLVMContext>());
 
 	for (const auto& unit : compilation_units) {
 		std::string module_name = get_module_name(unit);
-		codegen_visitor visitor(module_name);
+		module_holder mod_holder(module_name, *result->context, unit);
+
+		codegen_visitor visitor(*mod_holder.module);
 		try {
 			unit->accept(visitor);
 		}
@@ -24,18 +30,10 @@ std::vector<ir_holder> codegen::generate_ir() {
 			continue;
 		}
 
-		// extract IR from module
-		llvm::Module& module = visitor.get_module();
-		ir_holder holder;
-		holder.owning_unit = unit;
-		std::string ir_string;
-		llvm::raw_string_ostream rso(ir_string);
-		module.print(rso, nullptr);
-		holder.ir = rso.str();
-		ir_modules.push_back(holder);
+		result->modules.push_back(std::move(mod_holder));
 	}
 
-	return ir_modules;
+	return result;
 }
 
 std::string codegen::get_module_name(std::shared_ptr<compilation_unit> unit) {